removed adding login shit [][]][][][]

+++ ANDROID PSD/PSD/app/src/main/java/anon/psd/background/PSDServiceWorker.java
@@ -134,7 +134,7 @@ public abstract class PSDServiceWorker
     private void sendPassToService(PassItem pass)
     {
         Bundle bundle = new Bundle();
-        bundle.putInt("pass_item_id", pass.Id);
+        bundle.putShort("pass_item_id", pass.Id);
         Message msg = Message.obtain(null, MessageType.PasswordId.getInt(), bundle);
         sendMessage(msg);
         Log.d(TAG, "ServiceWorker Sent [ PASS ] command");
